requirements.txt
opencv-python
numpy
colour-science